tests/check/: Add Sparc ABI checks
Original commit message from CVS:
* tests/check/Makefile.am:
* tests/check/gst/gstabi.c:
* tests/check/gst/struct_sparc.h:
* tests/check/libs/libsabi.c:
* tests/check/libs/struct_sparc.h:
Add Sparc ABI checks
* tests/check/gst/gstvalue.c: (GST_START_TEST):
Cast signed integer to unsigned to avoid a compiler warning.

+++ tests/check/gst/gstvalue.c
@@ -139,7 +139,7 @@ GST_START_TEST (test_deserialize_guint64)
     0xffffffffffffffffULL,
     9223372036854775810ULL,
     9223372036854775806ULL,
-    -1,
+    (guint64) - 1,
     1,
     0,
   };
